finished (for now) user click information
made it so that the last row and column actually work and added a system to detect the coordinates of a users' mouse cursor when over the board. Trying next to make an onclick system in the pieces functions that, when a piece is clicked, log the coordinates of the piece and which piece was clicked to the console. After that, should be pretty simple to move pieces!