pieces know who they are!
bit of a confusing title, but basically the if blocks (they will most likely change to switches later) now know, once the user has clicked on a piece, which piece the user has clicked on. which is good! now, based on which piece is clicked, we can hopefully redraw the piece somewhere else and have it still be the same piece with all the functionality. Here's hoping!